feat: 수강신청만을 위한 클래스 Enrollment 추가
