+++ src/main/java/nextstep/session/domain/FreeSession.java
@@ -3,14 +3,8 @@ package nextstep.session.domain;
 import nextstep.courses.domain.Course;
 
 public class FreeSession extends Session {
-
 	public FreeSession(Course course, ImageInfo imageInfo, Period period, int maximumNumberOfParticipants) {
 		super(course, imageInfo, period, maximumNumberOfParticipants);
-	}
-
-	@Override
-	public boolean isMaximumNumberOfParticipantsLimited(int NumberOfParticipants) {
-		return true;
 	}
 
 	@Override
